+++ src/main/java/com/zz/pojo/HouseKeeper.java
@@ -5,14 +5,13 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Date;
 
 
 @Data
 public class HouseKeeper implements Serializable {
     @TableId(value = "house_keeper_id", type = IdType.ASSIGN_ID)
-    private BigInteger houseKeepingId;
+    private Integer houseKeepingId;
     private String houseKeeperName;
     private String password;
     private String companyName;
