+++ src/main/java/com/zz/pojo/Manager.java
@@ -5,12 +5,11 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 @Data
 public class Manager implements Serializable {
     @TableId(value = "manager_id", type = IdType.ASSIGN_ID)
-    private BigInteger managerId;
+    private Integer managerId;
     private String managerName;
     private String password;
